Desktop/x/src/back/requirements.txt
flask==3.0.2
flask_sqlalchemy==3.1.1
psycopg[binary]>=3.2.2
redis==5.0.1
flask-jwt-extended==4.6.0
flask-cors==4.0.0
bcrypt==4.1.2
python-dotenv==1.0.1
flask-migrate==4.0.5